requirements.txt
# Web scraping and HTTP
requests>=2.31.0
beautifulsoup4>=4.12.0

# AI and Machine Learning
google-generativeai==0.3.2

# Text-to-Speech and Translation
gTTS==2.5.0
googletrans==4.0.0-rc1

# Web Framework and API
fastapi==0.108.0
uvicorn==0.25.0
streamlit==1.29.0

# Data handling
pandas>=2.1.3
python-dotenv>=1.0.0

# Additional dependencies
plotly>=5.18.0
pydantic>=2.0.0
starlette>=0.27.0
python-multipart>=0.0.6
httpx>=0.24.1 